format decision in the battle bot module
